Add clippy CI, fix or ignore warnings where appropriate
