+++ simulator/runner/mod.rs
@@ -1,4 +1,5 @@
 pub mod cli;
 pub mod env;
+#[allow(dead_code)]
 pub mod file;
 pub mod io;
